requirements.txt
discord.py==2.3.2
deep-translator==1.11.4
langdetect==1.0.9
redis==5.0.1
aiosqlite==0.19.0
gTTS==2.4.0
Pillow==10.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
